pyproject.toml
[tool.poetry]
name = "bgt"
version = "0.1.0"
description = ""
authors = ["Tyler Peckenpaugh <tyler@ty-pe.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.45.0"
uvicorn = "^0.10.8"
boardgamegeek2 = "^1.0.1"
jsons = "^1.1.0"
boto3 = "^1.10.40"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.750"
autopep8 = "^1.4.4"
untangle = "^1.1.1"
xmltodict = "^0.12.0"
moto = "^1.3.14"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

